feat: Add new instructions.
This commit introduces a small number of implemented instructions.

Each instruction can now keep an array of argument objects. These argument objects are used to grab the value of the given argument (using `id`) from the opcode.

It also adds a `throw()` method for throwing errors (using the current state of the CPU as part of the error).

A rudimentary implementation of the CPU's display was also added.