feat: Add draw instruction.
This commit adds the `DRW Vx,Vy,N` instruction. THis required refactoring the existing implementation for the display inside the CPU. The display is now treated as one large array of 0 or 1 (off or on). Using a bit of math, it can be formatted as a 64x32 display.

It also adds some constants for the CPU. This is just `SCREEN_WIDTH` and `SCREEN_HEIGHT` for now.